rattler_config::config

Module concurrency

Structs

ConcurrencyConfig

Functions

default_max_concurrent_downloads

The default maximum number of concurrent downloads that can be run at once. 50 is a reasonable default for the number of concurrent downloads. More verification is needed to determine the optimal number.

default_max_concurrent_solves

The default maximum number of concurrent solves that can be run at once. Defaulting to the number of CPUs available.